[WIP] Refactor Apis to support multiple adapters
AuthApi now supports multiple adapters but is untested

TODO:
* Still need to migrate ModelApi and subclasses to new paradigm
* Implement the GraphQL adapters with ApolloClient
