+++ src/auth/actions/index.ts
@@ -1,4 +1,4 @@
-export * from "./AuthApi";
+export * from "./api/AuthApi";
 export * from "./creators";
 export * from "./thunks";
 export * from "./types";
