+++ src/auth/actions/types.ts
@@ -2,7 +2,7 @@ import { Action } from "redux";
 import { ThunkDispatch } from "redux-thunk";
 
 import { IUser } from "../../app/types";
-import { IAuthState } from "../model";
+import { IAuthState, IAuthInnerState } from "../model";
 
 const typeCache: { [label: string]: boolean } = {};
 
@@ -45,3 +45,9 @@ export interface ISetRedirect extends Action {
 
 export declare type AuthAction = Action | ICompleteLogin | ISuccessfulGetUserDetails | IFailLogin | ISetRedirect;
 export declare type AuthDispatch = ThunkDispatch<IAuthState, void, AuthAction>;
+
+export interface IAuthApiAdapter {
+  login: (username: string, password: string, dispatch: AuthDispatch, authState: IAuthInnerState) => Promise<any>;
+  socialLogin: (provider: string, code: string, redirectUri: string, dispatch: AuthDispatch, authState: IAuthInnerState) => Promise<any>;
+  logout: (dispatch: AuthDispatch, authState: IAuthInnerState) => Promise<any>;
+}
