+++ src/auth/pages/LogoutPage.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { connect } from "react-redux";
 import { RouteComponentProps, withRouter } from "react-router-dom";
 
-import { AuthApi, AuthDispatch as Dispatch } from "../actions";
+import { AuthApi, AuthDispatch as Dispatch, IAuthApiAdapter } from "../actions";
 
 export interface ILogoutPageDispatchProps {
   onLogout: () => void;
@@ -14,24 +14,26 @@ export interface ILogoutPageOwnProps extends RouteComponentProps<any> {
 
 export interface ILogoutPageProps extends ILogoutPageOwnProps, ILogoutPageDispatchProps {}
 
-class LogoutPage extends React.Component<ILogoutPageProps, {}> {
-  public componentWillMount() {
-    this.props.onLogout();
+export function buildLogoutPage(api: AuthApi) {
+  class LogoutPage extends React.Component<ILogoutPageProps, {}> {
+    public componentWillMount() {
+      this.props.onLogout();
+    }
+  
+    public render(): JSX.Element {
+      return null;
+    }
   }
-
-  public render(): JSX.Element {
-    return null;
+  
+  function mapDispatchToProps(dispatch: Dispatch, ownProps: ILogoutPageOwnProps): ILogoutPageDispatchProps {
+    return {
+      onLogout: () => {
+        dispatch(api.logout());
+        const redirectUri = ownProps.redirectUri ? ownProps.redirectUri : "/";
+        ownProps.history.replace(redirectUri);
+      },
+    };
   }
+  
+  return withRouter(connect(null, mapDispatchToProps)(LogoutPage));
 }
-
-function mapDispatchToProps(dispatch: Dispatch, ownProps: ILogoutPageOwnProps): ILogoutPageDispatchProps {
-  return {
-    onLogout: () => {
-      dispatch(AuthApi.getInstance().logout());
-      const redirectUri = ownProps.redirectUri ? ownProps.redirectUri : "/";
-      ownProps.history.replace(redirectUri);
-    },
-  };
-}
-
-export default withRouter(connect(null, mapDispatchToProps)(LogoutPage));
